requirements.txt
# Core ML and Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Deep Learning (optional)
torch>=1.12.0
tensorflow>=2.8.0

# Weights & Biases
wandb>=0.13.0

# Data Processing and Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter and Development
jupyter>=1.0.0
ipykernel>=6.0.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
tqdm>=4.64.0

# Additional ML utilities
imbalanced-learn>=0.9.0
optuna>=3.0.0 